@@ -35,6 +26,807 @@
       "./helpers/OverloadYield.js"
     ],
     "./helpers/esm/OverloadYield": "./helpers/esm/OverloadYield.js",
+    "./helpers/applyDecoratedDescriptor": [
+      {
+        "node": "./helpers/applyDecoratedDescriptor.js",
+        "import": "./helpers/esm/applyDecoratedDescriptor.js",
+        "default": "./helpers/applyDecoratedDescriptor.js"
+      },
+      "./helpers/applyDecoratedDescriptor.js"
+    ],
+    "./helpers/esm/applyDecoratedDescriptor": "./helpers/esm/applyDecoratedDescriptor.js",
+    "./helpers/applyDecs2311": [
+      {
+        "node": "./helpers/applyDecs2311.js",
+        "import": "./helpers/esm/applyDecs2311.js",
+        "default": "./helpers/applyDecs2311.js"
+      },
+      "./helpers/applyDecs2311.js"
+    ],
+    "./helpers/esm/applyDecs2311": "./helpers/esm/applyDecs2311.js",
+    "./helpers/arrayLikeToArray": [
+      {
+        "node": "./helpers/arrayLikeToArray.js",
+        "import": "./helpers/esm/arrayLikeToArray.js",
+        "default": "./helpers/arrayLikeToArray.js"
+      },
+      "./helpers/arrayLikeToArray.js"
+    ],
+    "./helpers/esm/arrayLikeToArray": "./helpers/esm/arrayLikeToArray.js",
+    "./helpers/arrayWithHoles": [
+      {
+        "node": "./helpers/arrayWithHoles.js",
+        "import": "./helpers/esm/arrayWithHoles.js",
+        "default": "./helpers/arrayWithHoles.js"
+      },
+      "./helpers/arrayWithHoles.js"
+    ],
+    "./helpers/esm/arrayWithHoles": "./helpers/esm/arrayWithHoles.js",
+    "./helpers/arrayWithoutHoles": [
+      {
+        "node": "./helpers/arrayWithoutHoles.js",
+        "import": "./helpers/esm/arrayWithoutHoles.js",
+        "default": "./helpers/arrayWithoutHoles.js"
+      },
+      "./helpers/arrayWithoutHoles.js"
+    ],
+    "./helpers/esm/arrayWithoutHoles": "./helpers/esm/arrayWithoutHoles.js",
+    "./helpers/assertClassBrand": [
+      {
+        "node": "./helpers/assertClassBrand.js",
+        "import": "./helpers/esm/assertClassBrand.js",
+        "default": "./helpers/assertClassBrand.js"
+      },
+      "./helpers/assertClassBrand.js"
+    ],
+    "./helpers/esm/assertClassBrand": "./helpers/esm/assertClassBrand.js",
+    "./helpers/assertThisInitialized": [
+      {
+        "node": "./helpers/assertThisInitialized.js",
+        "import": "./helpers/esm/assertThisInitialized.js",
+        "default": "./helpers/assertThisInitialized.js"
+      },
+      "./helpers/assertThisInitialized.js"
+    ],
+    "./helpers/esm/assertThisInitialized": "./helpers/esm/assertThisInitialized.js",
+    "./helpers/asyncGeneratorDelegate": [
+      {
+        "node": "./helpers/asyncGeneratorDelegate.js",
+        "import": "./helpers/esm/asyncGeneratorDelegate.js",
+        "default": "./helpers/asyncGeneratorDelegate.js"
+      },
+      "./helpers/asyncGeneratorDelegate.js"
+    ],
+    "./helpers/esm/asyncGeneratorDelegate": "./helpers/esm/asyncGeneratorDelegate.js",
+    "./helpers/asyncIterator": [
+      {
+        "node": "./helpers/asyncIterator.js",
+        "import": "./helpers/esm/asyncIterator.js",
+        "default": "./helpers/asyncIterator.js"
+      },
+      "./helpers/asyncIterator.js"
+    ],
+    "./helpers/esm/asyncIterator": "./helpers/esm/asyncIterator.js",
+    "./helpers/asyncToGenerator": [
+      {
+        "node": "./helpers/asyncToGenerator.js",
+        "import": "./helpers/esm/asyncToGenerator.js",
+        "default": "./helpers/asyncToGenerator.js"
+      },
+      "./helpers/asyncToGenerator.js"
+    ],
+    "./helpers/esm/asyncToGenerator": "./helpers/esm/asyncToGenerator.js",
+    "./helpers/awaitAsyncGenerator": [
+      {
+        "node": "./helpers/awaitAsyncGenerator.js",
+        "import": "./helpers/esm/awaitAsyncGenerator.js",
+        "default": "./helpers/awaitAsyncGenerator.js"
+      },
+      "./helpers/awaitAsyncGenerator.js"
+    ],
+    "./helpers/esm/awaitAsyncGenerator": "./helpers/esm/awaitAsyncGenerator.js",
+    "./helpers/callSuper": [
+      {
+        "node": "./helpers/callSuper.js",
+        "import": "./helpers/esm/callSuper.js",
+        "default": "./helpers/callSuper.js"
+      },
+      "./helpers/callSuper.js"
+    ],
+    "./helpers/esm/callSuper": "./helpers/esm/callSuper.js",
+    "./helpers/checkInRHS": [
+      {
+        "node": "./helpers/checkInRHS.js",
+        "import": "./helpers/esm/checkInRHS.js",
+        "default": "./helpers/checkInRHS.js"
+      },
+      "./helpers/checkInRHS.js"
+    ],
+    "./helpers/esm/checkInRHS": "./helpers/esm/checkInRHS.js",
+    "./helpers/checkPrivateRedeclaration": [
+      {
+        "node": "./helpers/checkPrivateRedeclaration.js",
+        "import": "./helpers/esm/checkPrivateRedeclaration.js",
+        "default": "./helpers/checkPrivateRedeclaration.js"
+      },
+      "./helpers/checkPrivateRedeclaration.js"
+    ],
+    "./helpers/esm/checkPrivateRedeclaration": "./helpers/esm/checkPrivateRedeclaration.js",
+    "./helpers/classCallCheck": [
+      {
+        "node": "./helpers/classCallCheck.js",
+        "import": "./helpers/esm/classCallCheck.js",
+        "default": "./helpers/classCallCheck.js"
+      },
+      "./helpers/classCallCheck.js"
+    ],
+    "./helpers/esm/classCallCheck": "./helpers/esm/classCallCheck.js",
+    "./helpers/classNameTDZError": [
+      {
+        "node": "./helpers/classNameTDZError.js",
+        "import": "./helpers/esm/classNameTDZError.js",
+        "default": "./helpers/classNameTDZError.js"
+      },
+      "./helpers/classNameTDZError.js"
+    ],
+    "./helpers/esm/classNameTDZError": "./helpers/esm/classNameTDZError.js",
+    "./helpers/classPrivateFieldGet2": [
+      {
+        "node": "./helpers/classPrivateFieldGet2.js",
+        "import": "./helpers/esm/classPrivateFieldGet2.js",
+        "default": "./helpers/classPrivateFieldGet2.js"
+      },
+      "./helpers/classPrivateFieldGet2.js"
+    ],
+    "./helpers/esm/classPrivateFieldGet2": "./helpers/esm/classPrivateFieldGet2.js",
+    "./helpers/classPrivateFieldInitSpec": [
+      {
+        "node": "./helpers/classPrivateFieldInitSpec.js",
+        "import": "./helpers/esm/classPrivateFieldInitSpec.js",
+        "default": "./helpers/classPrivateFieldInitSpec.js"
+      },
+      "./helpers/classPrivateFieldInitSpec.js"
+    ],
+    "./helpers/esm/classPrivateFieldInitSpec": "./helpers/esm/classPrivateFieldInitSpec.js",
+    "./helpers/classPrivateFieldLooseBase": [
+      {
+        "node": "./helpers/classPrivateFieldLooseBase.js",
+        "import": "./helpers/esm/classPrivateFieldLooseBase.js",
+        "default": "./helpers/classPrivateFieldLooseBase.js"
+      },
+      "./helpers/classPrivateFieldLooseBase.js"
+    ],
+    "./helpers/esm/classPrivateFieldLooseBase": "./helpers/esm/classPrivateFieldLooseBase.js",
+    "./helpers/classPrivateFieldLooseKey": [
+      {
+        "node": "./helpers/classPrivateFieldLooseKey.js",
+        "import": "./helpers/esm/classPrivateFieldLooseKey.js",
+        "default": "./helpers/classPrivateFieldLooseKey.js"
+      },
+      "./helpers/classPrivateFieldLooseKey.js"
+    ],
+    "./helpers/esm/classPrivateFieldLooseKey": "./helpers/esm/classPrivateFieldLooseKey.js",
+    "./helpers/classPrivateFieldSet2": [
+      {
+        "node": "./helpers/classPrivateFieldSet2.js",
+        "import": "./helpers/esm/classPrivateFieldSet2.js",
+        "default": "./helpers/classPrivateFieldSet2.js"
+      },
+      "./helpers/classPrivateFieldSet2.js"
+    ],
+    "./helpers/esm/classPrivateFieldSet2": "./helpers/esm/classPrivateFieldSet2.js",
+    "./helpers/classPrivateGetter": [
+      {
+        "node": "./helpers/classPrivateGetter.js",
+        "import": "./helpers/esm/classPrivateGetter.js",
+        "default": "./helpers/classPrivateGetter.js"
+      },
+      "./helpers/classPrivateGetter.js"
+    ],
+    "./helpers/esm/classPrivateGetter": "./helpers/esm/classPrivateGetter.js",
+    "./helpers/classPrivateMethodInitSpec": [
+      {
+        "node": "./helpers/classPrivateMethodInitSpec.js",
+        "import": "./helpers/esm/classPrivateMethodInitSpec.js",
+        "default": "./helpers/classPrivateMethodInitSpec.js"
+      },
+      "./helpers/classPrivateMethodInitSpec.js"
+    ],
+    "./helpers/esm/classPrivateMethodInitSpec": "./helpers/esm/classPrivateMethodInitSpec.js",
+    "./helpers/classPrivateSetter": [
+      {
+        "node": "./helpers/classPrivateSetter.js",
+        "import": "./helpers/esm/classPrivateSetter.js",
+        "default": "./helpers/classPrivateSetter.js"
+      },
+      "./helpers/classPrivateSetter.js"
+    ],
+    "./helpers/esm/classPrivateSetter": "./helpers/esm/classPrivateSetter.js",
+    "./helpers/classStaticPrivateMethodGet": [
+      {
+        "node": "./helpers/classStaticPrivateMethodGet.js",
+        "import": "./helpers/esm/classStaticPrivateMethodGet.js",
+        "default": "./helpers/classStaticPrivateMethodGet.js"
+      },
+      "./helpers/classStaticPrivateMethodGet.js"
+    ],
+    "./helpers/esm/classStaticPrivateMethodGet": "./helpers/esm/classStaticPrivateMethodGet.js",
+    "./helpers/construct": [
+      {
+        "node": "./helpers/construct.js",
+        "import": "./helpers/esm/construct.js",
+        "default": "./helpers/construct.js"
+      },
+      "./helpers/construct.js"
+    ],
+    "./helpers/esm/construct": "./helpers/esm/construct.js",
+    "./helpers/createClass": [
+      {
+        "node": "./helpers/createClass.js",
+        "import": "./helpers/esm/createClass.js",
+        "default": "./helpers/createClass.js"
+      },
+      "./helpers/createClass.js"
+    ],
+    "./helpers/esm/createClass": "./helpers/esm/createClass.js",
+    "./helpers/createForOfIteratorHelper": [
+      {
+        "node": "./helpers/createForOfIteratorHelper.js",
+        "import": "./helpers/esm/createForOfIteratorHelper.js",
+        "default": "./helpers/createForOfIteratorHelper.js"
+      },
+      "./helpers/createForOfIteratorHelper.js"
+    ],
+    "./helpers/esm/createForOfIteratorHelper": "./helpers/esm/createForOfIteratorHelper.js",
+    "./helpers/createForOfIteratorHelperLoose": [
+      {
+        "node": "./helpers/createForOfIteratorHelperLoose.js",
+        "import": "./helpers/esm/createForOfIteratorHelperLoose.js",
+        "default": "./helpers/createForOfIteratorHelperLoose.js"
+      },
+      "./helpers/createForOfIteratorHelperLoose.js"
+    ],
+    "./helpers/esm/createForOfIteratorHelperLoose": "./helpers/esm/createForOfIteratorHelperLoose.js",
+    "./helpers/createSuper": [
+      {
+        "node": "./helpers/createSuper.js",
+        "import": "./helpers/esm/createSuper.js",
+        "default": "./helpers/createSuper.js"
+      },
+      "./helpers/createSuper.js"
+    ],
+    "./helpers/esm/createSuper": "./helpers/esm/createSuper.js",
+    "./helpers/decorate": [
+      {
+        "node": "./helpers/decorate.js",
+        "import": "./helpers/esm/decorate.js",
+        "default": "./helpers/decorate.js"
+      },
+      "./helpers/decorate.js"
+    ],
+    "./helpers/esm/decorate": "./helpers/esm/decorate.js",
+    "./helpers/defaults": [
+      {
+        "node": "./helpers/defaults.js",
+        "import": "./helpers/esm/defaults.js",
+        "default": "./helpers/defaults.js"
+      },
+      "./helpers/defaults.js"
+    ],
+    "./helpers/esm/defaults": "./helpers/esm/defaults.js",
+    "./helpers/defineAccessor": [
+      {
+        "node": "./helpers/defineAccessor.js",
+        "import": "./helpers/esm/defineAccessor.js",
+        "default": "./helpers/defineAccessor.js"
+      },
+      "./helpers/defineAccessor.js"
+    ],
+    "./helpers/esm/defineAccessor": "./helpers/esm/defineAccessor.js",
+    "./helpers/defineProperty": [
+      {
+        "node": "./helpers/defineProperty.js",
+        "import": "./helpers/esm/defineProperty.js",
+        "default": "./helpers/defineProperty.js"
+      },
+      "./helpers/defineProperty.js"
+    ],
+    "./helpers/esm/defineProperty": "./helpers/esm/defineProperty.js",
+    "./helpers/extends": [
+      {
+        "node": "./helpers/extends.js",
+        "import": "./helpers/esm/extends.js",
+        "default": "./helpers/extends.js"
+      },
+      "./helpers/extends.js"
+    ],
+    "./helpers/esm/extends": "./helpers/esm/extends.js",
+    "./helpers/get": [
+      {
+        "node": "./helpers/get.js",
+        "import": "./helpers/esm/get.js",
+        "default": "./helpers/get.js"
+      },
+      "./helpers/get.js"
+    ],
+    "./helpers/esm/get": "./helpers/esm/get.js",
+    "./helpers/getPrototypeOf": [
+      {
+        "node": "./helpers/getPrototypeOf.js",
+        "import": "./helpers/esm/getPrototypeOf.js",
+        "default": "./helpers/getPrototypeOf.js"
+      },
+      "./helpers/getPrototypeOf.js"
+    ],
+    "./helpers/esm/getPrototypeOf": "./helpers/esm/getPrototypeOf.js",
+    "./helpers/identity": [
+      {
+        "node": "./helpers/identity.js",
+        "import": "./helpers/esm/identity.js",
+        "default": "./helpers/identity.js"
+      },
+      "./helpers/identity.js"
+    ],
+    "./helpers/esm/identity": "./helpers/esm/identity.js",
+    "./helpers/importDeferProxy": [
+      {
+        "node": "./helpers/importDeferProxy.js",
+        "import": "./helpers/esm/importDeferProxy.js",
+        "default": "./helpers/importDeferProxy.js"
+      },
+      "./helpers/importDeferProxy.js"
+    ],
+    "./helpers/esm/importDeferProxy": "./helpers/esm/importDeferProxy.js",
+    "./helpers/inherits": [
+      {
+        "node": "./helpers/inherits.js",
+        "import": "./helpers/esm/inherits.js",
+        "default": "./helpers/inherits.js"
+      },
+      "./helpers/inherits.js"
+    ],
+    "./helpers/esm/inherits": "./helpers/esm/inherits.js",
+    "./helpers/inheritsLoose": [
+      {
+        "node": "./helpers/inheritsLoose.js",
+        "import": "./helpers/esm/inheritsLoose.js",
+        "default": "./helpers/inheritsLoose.js"
+      },
+      "./helpers/inheritsLoose.js"
+    ],
+    "./helpers/esm/inheritsLoose": "./helpers/esm/inheritsLoose.js",
+    "./helpers/initializerDefineProperty": [
+      {
+        "node": "./helpers/initializerDefineProperty.js",
+        "import": "./helpers/esm/initializerDefineProperty.js",
+        "default": "./helpers/initializerDefineProperty.js"
+      },
+      "./helpers/initializerDefineProperty.js"
+    ],
+    "./helpers/esm/initializerDefineProperty": "./helpers/esm/initializerDefineProperty.js",
+    "./helpers/initializerWarningHelper": [
+      {
+        "node": "./helpers/initializerWarningHelper.js",
+        "import": "./helpers/esm/initializerWarningHelper.js",
+        "default": "./helpers/initializerWarningHelper.js"
+      },
+      "./helpers/initializerWarningHelper.js"
+    ],
+    "./helpers/esm/initializerWarningHelper": "./helpers/esm/initializerWarningHelper.js",
+    "./helpers/instanceof": [
+      {
+        "node": "./helpers/instanceof.js",
+        "import": "./helpers/esm/instanceof.js",
+        "default": "./helpers/instanceof.js"
+      },
+      "./helpers/instanceof.js"
+    ],
+    "./helpers/esm/instanceof": "./helpers/esm/instanceof.js",
+    "./helpers/interopRequireDefault": [
+      {
+        "node": "./helpers/interopRequireDefault.js",
+        "import": "./helpers/esm/interopRequireDefault.js",
+        "default": "./helpers/interopRequireDefault.js"
+      },
+      "./helpers/interopRequireDefault.js"
+    ],
+    "./helpers/esm/interopRequireDefault": "./helpers/esm/interopRequireDefault.js",
+    "./helpers/interopRequireWildcard": [
+      {
+        "node": "./helpers/interopRequireWildcard.js",
+        "import": "./helpers/esm/interopRequireWildcard.js",
+        "default": "./helpers/interopRequireWildcard.js"
+      },
+      "./helpers/interopRequireWildcard.js"
+    ],
+    "./helpers/esm/interopRequireWildcard": "./helpers/esm/interopRequireWildcard.js",
+    "./helpers/isNativeFunction": [
+      {
+        "node": "./helpers/isNativeFunction.js",
+        "import": "./helpers/esm/isNativeFunction.js",
+        "default": "./helpers/isNativeFunction.js"
+      },
+      "./helpers/isNativeFunction.js"
+    ],
+    "./helpers/esm/isNativeFunction": "./helpers/esm/isNativeFunction.js",
+    "./helpers/isNativeReflectConstruct": [
+      {
+        "node": "./helpers/isNativeReflectConstruct.js",
+        "import": "./helpers/esm/isNativeReflectConstruct.js",
+        "default": "./helpers/isNativeReflectConstruct.js"
+      },
+      "./helpers/isNativeReflectConstruct.js"
+    ],
+    "./helpers/esm/isNativeReflectConstruct": "./helpers/esm/isNativeReflectConstruct.js",
+    "./helpers/iterableToArray": [
+      {
+        "node": "./helpers/iterableToArray.js",
+        "import": "./helpers/esm/iterableToArray.js",
+        "default": "./helpers/iterableToArray.js"
+      },
+      "./helpers/iterableToArray.js"
+    ],
+    "./helpers/esm/iterableToArray": "./helpers/esm/iterableToArray.js",
+    "./helpers/iterableToArrayLimit": [
+      {
+        "node": "./helpers/iterableToArrayLimit.js",
+        "import": "./helpers/esm/iterableToArrayLimit.js",
+        "default": "./helpers/iterableToArrayLimit.js"
+      },
+      "./helpers/iterableToArrayLimit.js"
+    ],
+    "./helpers/esm/iterableToArrayLimit": "./helpers/esm/iterableToArrayLimit.js",
+    "./helpers/jsx": [
+      {
+        "node": "./helpers/jsx.js",
+        "import": "./helpers/esm/jsx.js",
+        "default": "./helpers/jsx.js"
+      },
+      "./helpers/jsx.js"
+    ],
+    "./helpers/esm/jsx": "./helpers/esm/jsx.js",
+    "./helpers/maybeArrayLike": [
+      {
+        "node": "./helpers/maybeArrayLike.js",
+        "import": "./helpers/esm/maybeArrayLike.js",
+        "default": "./helpers/maybeArrayLike.js"
+      },
+      "./helpers/maybeArrayLike.js"
+    ],
+    "./helpers/esm/maybeArrayLike": "./helpers/esm/maybeArrayLike.js",
+    "./helpers/newArrowCheck": [
+      {
+        "node": "./helpers/newArrowCheck.js",
+        "import": "./helpers/esm/newArrowCheck.js",
+        "default": "./helpers/newArrowCheck.js"
+      },
+      "./helpers/newArrowCheck.js"
+    ],
+    "./helpers/esm/newArrowCheck": "./helpers/esm/newArrowCheck.js",
+    "./helpers/nonIterableRest": [
+      {
+        "node": "./helpers/nonIterableRest.js",
+        "import": "./helpers/esm/nonIterableRest.js",
+        "default": "./helpers/nonIterableRest.js"
+      },
+      "./helpers/nonIterableRest.js"
+    ],
+    "./helpers/esm/nonIterableRest": "./helpers/esm/nonIterableRest.js",
+    "./helpers/nonIterableSpread": [
+      {
+        "node": "./helpers/nonIterableSpread.js",
+        "import": "./helpers/esm/nonIterableSpread.js",
+        "default": "./helpers/nonIterableSpread.js"
+      },
+      "./helpers/nonIterableSpread.js"
+    ],
+    "./helpers/esm/nonIterableSpread": "./helpers/esm/nonIterableSpread.js",
+    "./helpers/nullishReceiverError": [
+      {
+        "node": "./helpers/nullishReceiverError.js",
+        "import": "./helpers/esm/nullishReceiverError.js",
+        "default": "./helpers/nullishReceiverError.js"
+      },
+      "./helpers/nullishReceiverError.js"
+    ],
+    "./helpers/esm/nullishReceiverError": "./helpers/esm/nullishReceiverError.js",
+    "./helpers/objectDestructuringEmpty": [
+      {
+        "node": "./helpers/objectDestructuringEmpty.js",
+        "import": "./helpers/esm/objectDestructuringEmpty.js",
+        "default": "./helpers/objectDestructuringEmpty.js"
+      },
+      "./helpers/objectDestructuringEmpty.js"
+    ],
+    "./helpers/esm/objectDestructuringEmpty": "./helpers/esm/objectDestructuringEmpty.js",
+    "./helpers/objectSpread2": [
+      {
+        "node": "./helpers/objectSpread2.js",
+        "import": "./helpers/esm/objectSpread2.js",
+        "default": "./helpers/objectSpread2.js"
+      },
+      "./helpers/objectSpread2.js"
+    ],
+    "./helpers/esm/objectSpread2": "./helpers/esm/objectSpread2.js",
+    "./helpers/objectWithoutProperties": [
+      {
+        "node": "./helpers/objectWithoutProperties.js",
+        "import": "./helpers/esm/objectWithoutProperties.js",
+        "default": "./helpers/objectWithoutProperties.js"
+      },
+      "./helpers/objectWithoutProperties.js"
+    ],
+    "./helpers/esm/objectWithoutProperties": "./helpers/esm/objectWithoutProperties.js",
+    "./helpers/objectWithoutPropertiesLoose": [
+      {
+        "node": "./helpers/objectWithoutPropertiesLoose.js",
+        "import": "./helpers/esm/objectWithoutPropertiesLoose.js",
+        "default": "./helpers/objectWithoutPropertiesLoose.js"
+      },
+      "./helpers/objectWithoutPropertiesLoose.js"
+    ],
+    "./helpers/esm/objectWithoutPropertiesLoose": "./helpers/esm/objectWithoutPropertiesLoose.js",
+    "./helpers/possibleConstructorReturn": [
+      {
+        "node": "./helpers/possibleConstructorReturn.js",
+        "import": "./helpers/esm/possibleConstructorReturn.js",
+        "default": "./helpers/possibleConstructorReturn.js"
+      },
+      "./helpers/possibleConstructorReturn.js"
+    ],
+    "./helpers/esm/possibleConstructorReturn": "./helpers/esm/possibleConstructorReturn.js",
+    "./helpers/readOnlyError": [
+      {
+        "node": "./helpers/readOnlyError.js",
+        "import": "./helpers/esm/readOnlyError.js",
+        "default": "./helpers/readOnlyError.js"
+      },
+      "./helpers/readOnlyError.js"
+    ],
+    "./helpers/esm/readOnlyError": "./helpers/esm/readOnlyError.js",
+    "./helpers/regeneratorRuntime": [
+      {
+        "node": "./helpers/regeneratorRuntime.js",
+        "import": "./helpers/esm/regeneratorRuntime.js",
+        "default": "./helpers/regeneratorRuntime.js"
+      },
+      "./helpers/regeneratorRuntime.js"
+    ],
+    "./helpers/esm/regeneratorRuntime": "./helpers/esm/regeneratorRuntime.js",
+    "./helpers/set": [
+      {
+        "node": "./helpers/set.js",
+        "import": "./helpers/esm/set.js",
+        "default": "./helpers/set.js"
+      },
+      "./helpers/set.js"
+    ],
+    "./helpers/esm/set": "./helpers/esm/set.js",
+    "./helpers/setFunctionName": [
+      {
+        "node": "./helpers/setFunctionName.js",
+        "import": "./helpers/esm/setFunctionName.js",
+        "default": "./helpers/setFunctionName.js"
+      },
+      "./helpers/setFunctionName.js"
+    ],
+    "./helpers/esm/setFunctionName": "./helpers/esm/setFunctionName.js",
+    "./helpers/setPrototypeOf": [
+      {
+        "node": "./helpers/setPrototypeOf.js",
+        "import": "./helpers/esm/setPrototypeOf.js",
+        "default": "./helpers/setPrototypeOf.js"
+      },
+      "./helpers/setPrototypeOf.js"
+    ],
+    "./helpers/esm/setPrototypeOf": "./helpers/esm/setPrototypeOf.js",
+    "./helpers/skipFirstGeneratorNext": [
+      {
+        "node": "./helpers/skipFirstGeneratorNext.js",
+        "import": "./helpers/esm/skipFirstGeneratorNext.js",
+        "default": "./helpers/skipFirstGeneratorNext.js"
+      },
+      "./helpers/skipFirstGeneratorNext.js"
+    ],
+    "./helpers/esm/skipFirstGeneratorNext": "./helpers/esm/skipFirstGeneratorNext.js",
+    "./helpers/slicedToArray": [
+      {
+        "node": "./helpers/slicedToArray.js",
+        "import": "./helpers/esm/slicedToArray.js",
+        "default": "./helpers/slicedToArray.js"
+      },
+      "./helpers/slicedToArray.js"
+    ],
+    "./helpers/esm/slicedToArray": "./helpers/esm/slicedToArray.js",
+    "./helpers/superPropBase": [
+      {
+        "node": "./helpers/superPropBase.js",
+        "import": "./helpers/esm/superPropBase.js",
+        "default": "./helpers/superPropBase.js"
+      },
+      "./helpers/superPropBase.js"
+    ],
+    "./helpers/esm/superPropBase": "./helpers/esm/superPropBase.js",
+    "./helpers/superPropGet": [
+      {
+        "node": "./helpers/superPropGet.js",
+        "import": "./helpers/esm/superPropGet.js",
+        "default": "./helpers/superPropGet.js"
+      },
+      "./helpers/superPropGet.js"
+    ],
+    "./helpers/esm/superPropGet": "./helpers/esm/superPropGet.js",
+    "./helpers/superPropSet": [
+      {
+        "node": "./helpers/superPropSet.js",
+        "import": "./helpers/esm/superPropSet.js",
+        "default": "./helpers/superPropSet.js"
+      },
+      "./helpers/superPropSet.js"
+    ],
+    "./helpers/esm/superPropSet": "./helpers/esm/superPropSet.js",
+    "./helpers/taggedTemplateLiteral": [
+      {
+        "node": "./helpers/taggedTemplateLiteral.js",
+        "import": "./helpers/esm/taggedTemplateLiteral.js",
+        "default": "./helpers/taggedTemplateLiteral.js"
+      },
+      "./helpers/taggedTemplateLiteral.js"
+    ],
+    "./helpers/esm/taggedTemplateLiteral": "./helpers/esm/taggedTemplateLiteral.js",
+    "./helpers/taggedTemplateLiteralLoose": [
+      {
+        "node": "./helpers/taggedTemplateLiteralLoose.js",
+        "import": "./helpers/esm/taggedTemplateLiteralLoose.js",
+        "default": "./helpers/taggedTemplateLiteralLoose.js"
+      },
+      "./helpers/taggedTemplateLiteralLoose.js"
+    ],
+    "./helpers/esm/taggedTemplateLiteralLoose": "./helpers/esm/taggedTemplateLiteralLoose.js",
+    "./helpers/tdz": [
+      {
+        "node": "./helpers/tdz.js",
+        "import": "./helpers/esm/tdz.js",
+        "default": "./helpers/tdz.js"
+      },
+      "./helpers/tdz.js"
+    ],
+    "./helpers/esm/tdz": "./helpers/esm/tdz.js",
+    "./helpers/temporalRef": [
+      {
+        "node": "./helpers/temporalRef.js",
+        "import": "./helpers/esm/temporalRef.js",
+        "default": "./helpers/temporalRef.js"
+      },
+      "./helpers/temporalRef.js"
+    ],
+    "./helpers/esm/temporalRef": "./helpers/esm/temporalRef.js",
+    "./helpers/temporalUndefined": [
+      {
+        "node": "./helpers/temporalUndefined.js",
+        "import": "./helpers/esm/temporalUndefined.js",
+        "default": "./helpers/temporalUndefined.js"
+      },
+      "./helpers/temporalUndefined.js"
+    ],
+    "./helpers/esm/temporalUndefined": "./helpers/esm/temporalUndefined.js",
+    "./helpers/toArray": [
+      {
+        "node": "./helpers/toArray.js",
+        "import": "./helpers/esm/toArray.js",
+        "default": "./helpers/toArray.js"
+      },
+      "./helpers/toArray.js"
+    ],
+    "./helpers/esm/toArray": "./helpers/esm/toArray.js",
+    "./helpers/toConsumableArray": [
+      {
+        "node": "./helpers/toConsumableArray.js",
+        "import": "./helpers/esm/toConsumableArray.js",
+        "default": "./helpers/toConsumableArray.js"
+      },
+      "./helpers/toConsumableArray.js"
+    ],
+    "./helpers/esm/toConsumableArray": "./helpers/esm/toConsumableArray.js",
+    "./helpers/toPrimitive": [
+      {
+        "node": "./helpers/toPrimitive.js",
+        "import": "./helpers/esm/toPrimitive.js",
+        "default": "./helpers/toPrimitive.js"
+      },
+      "./helpers/toPrimitive.js"
+    ],
+    "./helpers/esm/toPrimitive": "./helpers/esm/toPrimitive.js",
+    "./helpers/toPropertyKey": [
+      {
+        "node": "./helpers/toPropertyKey.js",
+        "import": "./helpers/esm/toPropertyKey.js",
+        "default": "./helpers/toPropertyKey.js"
+      },
+      "./helpers/toPropertyKey.js"
+    ],
+    "./helpers/esm/toPropertyKey": "./helpers/esm/toPropertyKey.js",
+    "./helpers/toSetter": [
+      {
+        "node": "./helpers/toSetter.js",
+        "import": "./helpers/esm/toSetter.js",
+        "default": "./helpers/toSetter.js"
+      },
+      "./helpers/toSetter.js"
+    ],
+    "./helpers/esm/toSetter": "./helpers/esm/toSetter.js",
+    "./helpers/typeof": [
+      {
+        "node": "./helpers/typeof.js",
+        "import": "./helpers/esm/typeof.js",
+        "default": "./helpers/typeof.js"
+      },
+      "./helpers/typeof.js"
+    ],
+    "./helpers/esm/typeof": "./helpers/esm/typeof.js",
+    "./helpers/unsupportedIterableToArray": [
+      {
+        "node": "./helpers/unsupportedIterableToArray.js",
+        "import": "./helpers/esm/unsupportedIterableToArray.js",
+        "default": "./helpers/unsupportedIterableToArray.js"
+      },
+      "./helpers/unsupportedIterableToArray.js"
+    ],
+    "./helpers/esm/unsupportedIterableToArray": "./helpers/esm/unsupportedIterableToArray.js",
+    "./helpers/usingCtx": [
+      {
+        "node": "./helpers/usingCtx.js",
+        "import": "./helpers/esm/usingCtx.js",
+        "default": "./helpers/usingCtx.js"
+      },
+      "./helpers/usingCtx.js"
+    ],
+    "./helpers/esm/usingCtx": "./helpers/esm/usingCtx.js",
+    "./helpers/wrapAsyncGenerator": [
+      {
+        "node": "./helpers/wrapAsyncGenerator.js",
+        "import": "./helpers/esm/wrapAsyncGenerator.js",
+        "default": "./helpers/wrapAsyncGenerator.js"
+      },
+      "./helpers/wrapAsyncGenerator.js"
+    ],
+    "./helpers/esm/wrapAsyncGenerator": "./helpers/esm/wrapAsyncGenerator.js",
+    "./helpers/wrapNativeSuper": [
+      {
+        "node": "./helpers/wrapNativeSuper.js",
+        "import": "./helpers/esm/wrapNativeSuper.js",
+        "default": "./helpers/wrapNativeSuper.js"
+      },
+      "./helpers/wrapNativeSuper.js"
+    ],
+    "./helpers/esm/wrapNativeSuper": "./helpers/esm/wrapNativeSuper.js",
+    "./helpers/wrapRegExp": [
+      {
+        "node": "./helpers/wrapRegExp.js",
+        "import": "./helpers/esm/wrapRegExp.js",
+        "default": "./helpers/wrapRegExp.js"
+      },
+      "./helpers/wrapRegExp.js"
+    ],
+    "./helpers/esm/wrapRegExp": "./helpers/esm/wrapRegExp.js",
+    "./helpers/writeOnlyError": [
+      {
+        "node": "./helpers/writeOnlyError.js",
+        "import": "./helpers/esm/writeOnlyError.js",
+        "default": "./helpers/writeOnlyError.js"
+      },
+      "./helpers/writeOnlyError.js"
+    ],
+    "./helpers/esm/writeOnlyError": "./helpers/esm/writeOnlyError.js",
+    "./helpers/AwaitValue": [
+      {
+        "node": "./helpers/AwaitValue.js",
+        "import": "./helpers/esm/AwaitValue.js",
+        "default": "./helpers/AwaitValue.js"
+      },
+      "./helpers/AwaitValue.js"
+    ],
+    "./helpers/esm/AwaitValue": "./helpers/esm/AwaitValue.js",
     "./helpers/applyDecs": [
       {
         "node": "./helpers/applyDecs.js",
@@ -71,735 +863,24 @@
       "./helpers/applyDecs2301.js"
     ],
     "./helpers/esm/applyDecs2301": "./helpers/esm/applyDecs2301.js",
-    "./helpers/asyncGeneratorDelegate": [
+    "./helpers/applyDecs2305": [
       {
-        "node": "./helpers/asyncGeneratorDelegate.js",
-        "import": "./helpers/esm/asyncGeneratorDelegate.js",
-        "default": "./helpers/asyncGeneratorDelegate.js"
+        "node": "./helpers/applyDecs2305.js",
+        "import": "./helpers/esm/applyDecs2305.js",
+        "default": "./helpers/applyDecs2305.js"
       },
-      "./helpers/asyncGeneratorDelegate.js"
+      "./helpers/applyDecs2305.js"
     ],
-    "./helpers/esm/asyncGeneratorDelegate": "./helpers/esm/asyncGeneratorDelegate.js",
-    "./helpers/asyncIterator": [
+    "./helpers/esm/applyDecs2305": "./helpers/esm/applyDecs2305.js",
+    "./helpers/classApplyDescriptorDestructureSet": [
       {
-        "node": "./helpers/asyncIterator.js",
-        "import": "./helpers/esm/asyncIterator.js",
-        "default": "./helpers/asyncIterator.js"
+        "node": "./helpers/classApplyDescriptorDestructureSet.js",
+        "import": "./helpers/esm/classApplyDescriptorDestructureSet.js",
+        "default": "./helpers/classApplyDescriptorDestructureSet.js"
       },
-      "./helpers/asyncIterator.js"
+      "./helpers/classApplyDescriptorDestructureSet.js"
     ],
-    "./helpers/esm/asyncIterator": "./helpers/esm/asyncIterator.js",
-    "./helpers/awaitAsyncGenerator": [
-      {
-        "node": "./helpers/awaitAsyncGenerator.js",
-        "import": "./helpers/esm/awaitAsyncGenerator.js",
-        "default": "./helpers/awaitAsyncGenerator.js"
-      },
-      "./helpers/awaitAsyncGenerator.js"
-    ],
-    "./helpers/esm/awaitAsyncGenerator": "./helpers/esm/awaitAsyncGenerator.js",
-    "./helpers/checkInRHS": [
-      {
-        "node": "./helpers/checkInRHS.js",
-        "import": "./helpers/esm/checkInRHS.js",
-        "default": "./helpers/checkInRHS.js"
-      },
-      "./helpers/checkInRHS.js"
-    ],
-    "./helpers/esm/checkInRHS": "./helpers/esm/checkInRHS.js",
-    "./helpers/defineAccessor": [
-      {
-        "node": "./helpers/defineAccessor.js",
-        "import": "./helpers/esm/defineAccessor.js",
-        "default": "./helpers/defineAccessor.js"
-      },
-      "./helpers/defineAccessor.js"
-    ],
-    "./helpers/esm/defineAccessor": "./helpers/esm/defineAccessor.js",
-    "./helpers/iterableToArrayLimit": [
-      {
-        "node": "./helpers/iterableToArrayLimit.js",
-        "import": "./helpers/esm/iterableToArrayLimit.js",
-        "default": "./helpers/iterableToArrayLimit.js"
-      },
-      "./helpers/iterableToArrayLimit.js"
-    ],
-    "./helpers/esm/iterableToArrayLimit": "./helpers/esm/iterableToArrayLimit.js",
-    "./helpers/iterableToArrayLimitLoose": [
-      {
-        "node": "./helpers/iterableToArrayLimitLoose.js",
-        "import": "./helpers/esm/iterableToArrayLimitLoose.js",
-        "default": "./helpers/iterableToArrayLimitLoose.js"
-      },
-      "./helpers/iterableToArrayLimitLoose.js"
-    ],
-    "./helpers/esm/iterableToArrayLimitLoose": "./helpers/esm/iterableToArrayLimitLoose.js",
-    "./helpers/jsx": [
-      {
-        "node": "./helpers/jsx.js",
-        "import": "./helpers/esm/jsx.js",
-        "default": "./helpers/jsx.js"
-      },
-      "./helpers/jsx.js"
-    ],
-    "./helpers/esm/jsx": "./helpers/esm/jsx.js",
-    "./helpers/objectSpread2": [
-      {
-        "node": "./helpers/objectSpread2.js",
-        "import": "./helpers/esm/objectSpread2.js",
-        "default": "./helpers/objectSpread2.js"
-      },
-      "./helpers/objectSpread2.js"
-    ],
-    "./helpers/esm/objectSpread2": "./helpers/esm/objectSpread2.js",
-    "./helpers/regeneratorRuntime": [
-      {
-        "node": "./helpers/regeneratorRuntime.js",
-        "import": "./helpers/esm/regeneratorRuntime.js",
-        "default": "./helpers/regeneratorRuntime.js"
-      },
-      "./helpers/regeneratorRuntime.js"
-    ],
-    "./helpers/esm/regeneratorRuntime": "./helpers/esm/regeneratorRuntime.js",
-    "./helpers/typeof": [
-      {
-        "node": "./helpers/typeof.js",
-        "import": "./helpers/esm/typeof.js",
-        "default": "./helpers/typeof.js"
-      },
-      "./helpers/typeof.js"
-    ],
-    "./helpers/esm/typeof": "./helpers/esm/typeof.js",
-    "./helpers/wrapRegExp": [
-      {
-        "node": "./helpers/wrapRegExp.js",
-        "import": "./helpers/esm/wrapRegExp.js",
-        "default": "./helpers/wrapRegExp.js"
-      },
-      "./helpers/wrapRegExp.js"
-    ],
-    "./helpers/esm/wrapRegExp": "./helpers/esm/wrapRegExp.js",
-    "./helpers/AwaitValue": [
-      {
-        "node": "./helpers/AwaitValue.js",
-        "import": "./helpers/esm/AwaitValue.js",
-        "default": "./helpers/AwaitValue.js"
-      },
-      "./helpers/AwaitValue.js"
-    ],
-    "./helpers/esm/AwaitValue": "./helpers/esm/AwaitValue.js",
-    "./helpers/wrapAsyncGenerator": [
-      {
-        "node": "./helpers/wrapAsyncGenerator.js",
-        "import": "./helpers/esm/wrapAsyncGenerator.js",
-        "default": "./helpers/wrapAsyncGenerator.js"
-      },
-      "./helpers/wrapAsyncGenerator.js"
-    ],
-    "./helpers/esm/wrapAsyncGenerator": "./helpers/esm/wrapAsyncGenerator.js",
-    "./helpers/asyncToGenerator": [
-      {
-        "node": "./helpers/asyncToGenerator.js",
-        "import": "./helpers/esm/asyncToGenerator.js",
-        "default": "./helpers/asyncToGenerator.js"
-      },
-      "./helpers/asyncToGenerator.js"
-    ],
-    "./helpers/esm/asyncToGenerator": "./helpers/esm/asyncToGenerator.js",
-    "./helpers/classCallCheck": [
-      {
-        "node": "./helpers/classCallCheck.js",
-        "import": "./helpers/esm/classCallCheck.js",
-        "default": "./helpers/classCallCheck.js"
-      },
-      "./helpers/classCallCheck.js"
-    ],
-    "./helpers/esm/classCallCheck": "./helpers/esm/classCallCheck.js",
-    "./helpers/createClass": [
-      {
-        "node": "./helpers/createClass.js",
-        "import": "./helpers/esm/createClass.js",
-        "default": "./helpers/createClass.js"
-      },
-      "./helpers/createClass.js"
-    ],
-    "./helpers/esm/createClass": "./helpers/esm/createClass.js",
-    "./helpers/defineEnumerableProperties": [
-      {
-        "node": "./helpers/defineEnumerableProperties.js",
-        "import": "./helpers/esm/defineEnumerableProperties.js",
-        "default": "./helpers/defineEnumerableProperties.js"
-      },
-      "./helpers/defineEnumerableProperties.js"
-    ],
-    "./helpers/esm/defineEnumerableProperties": "./helpers/esm/defineEnumerableProperties.js",
-    "./helpers/defaults": [
-      {
-        "node": "./helpers/defaults.js",
-        "import": "./helpers/esm/defaults.js",
-        "default": "./helpers/defaults.js"
-      },
-      "./helpers/defaults.js"
-    ],
-    "./helpers/esm/defaults": "./helpers/esm/defaults.js",
-    "./helpers/defineProperty": [
-      {
-        "node": "./helpers/defineProperty.js",
-        "import": "./helpers/esm/defineProperty.js",
-        "default": "./helpers/defineProperty.js"
-      },
-      "./helpers/defineProperty.js"
-    ],
-    "./helpers/esm/defineProperty": "./helpers/esm/defineProperty.js",
-    "./helpers/extends": [
-      {
-        "node": "./helpers/extends.js",
-        "import": "./helpers/esm/extends.js",
-        "default": "./helpers/extends.js"
-      },
-      "./helpers/extends.js"
-    ],
-    "./helpers/esm/extends": "./helpers/esm/extends.js",
-    "./helpers/objectSpread": [
-      {
-        "node": "./helpers/objectSpread.js",
-        "import": "./helpers/esm/objectSpread.js",
-        "default": "./helpers/objectSpread.js"
-      },
-      "./helpers/objectSpread.js"
-    ],
-    "./helpers/esm/objectSpread": "./helpers/esm/objectSpread.js",
-    "./helpers/inherits": [
-      {
-        "node": "./helpers/inherits.js",
-        "import": "./helpers/esm/inherits.js",
-        "default": "./helpers/inherits.js"
-      },
-      "./helpers/inherits.js"
-    ],
-    "./helpers/esm/inherits": "./helpers/esm/inherits.js",
-    "./helpers/inheritsLoose": [
-      {
-        "node": "./helpers/inheritsLoose.js",
-        "import": "./helpers/esm/inheritsLoose.js",
-        "default": "./helpers/inheritsLoose.js"
-      },
-      "./helpers/inheritsLoose.js"
-    ],
-    "./helpers/esm/inheritsLoose": "./helpers/esm/inheritsLoose.js",
-    "./helpers/getPrototypeOf": [
-      {
-        "node": "./helpers/getPrototypeOf.js",
-        "import": "./helpers/esm/getPrototypeOf.js",
-        "default": "./helpers/getPrototypeOf.js"
-      },
-      "./helpers/getPrototypeOf.js"
-    ],
-    "./helpers/esm/getPrototypeOf": "./helpers/esm/getPrototypeOf.js",
-    "./helpers/setPrototypeOf": [
-      {
-        "node": "./helpers/setPrototypeOf.js",
-        "import": "./helpers/esm/setPrototypeOf.js",
-        "default": "./helpers/setPrototypeOf.js"
-      },
-      "./helpers/setPrototypeOf.js"
-    ],
-    "./helpers/esm/setPrototypeOf": "./helpers/esm/setPrototypeOf.js",
-    "./helpers/isNativeReflectConstruct": [
-      {
-        "node": "./helpers/isNativeReflectConstruct.js",
-        "import": "./helpers/esm/isNativeReflectConstruct.js",
-        "default": "./helpers/isNativeReflectConstruct.js"
-      },
-      "./helpers/isNativeReflectConstruct.js"
-    ],
-    "./helpers/esm/isNativeReflectConstruct": "./helpers/esm/isNativeReflectConstruct.js",
-    "./helpers/construct": [
-      {
-        "node": "./helpers/construct.js",
-        "import": "./helpers/esm/construct.js",
-        "default": "./helpers/construct.js"
-      },
-      "./helpers/construct.js"
-    ],
-    "./helpers/esm/construct": "./helpers/esm/construct.js",
-    "./helpers/isNativeFunction": [
-      {
-        "node": "./helpers/isNativeFunction.js",
-        "import": "./helpers/esm/isNativeFunction.js",
-        "default": "./helpers/isNativeFunction.js"
-      },
-      "./helpers/isNativeFunction.js"
-    ],
-    "./helpers/esm/isNativeFunction": "./helpers/esm/isNativeFunction.js",
-    "./helpers/wrapNativeSuper": [
-      {
-        "node": "./helpers/wrapNativeSuper.js",
-        "import": "./helpers/esm/wrapNativeSuper.js",
-        "default": "./helpers/wrapNativeSuper.js"
-      },
-      "./helpers/wrapNativeSuper.js"
-    ],
-    "./helpers/esm/wrapNativeSuper": "./helpers/esm/wrapNativeSuper.js",
-    "./helpers/instanceof": [
-      {
-        "node": "./helpers/instanceof.js",
-        "import": "./helpers/esm/instanceof.js",
-        "default": "./helpers/instanceof.js"
-      },
-      "./helpers/instanceof.js"
-    ],
-    "./helpers/esm/instanceof": "./helpers/esm/instanceof.js",
-    "./helpers/interopRequireDefault": [
-      {
-        "node": "./helpers/interopRequireDefault.js",
-        "import": "./helpers/esm/interopRequireDefault.js",
-        "default": "./helpers/interopRequireDefault.js"
-      },
-      "./helpers/interopRequireDefault.js"
-    ],
-    "./helpers/esm/interopRequireDefault": "./helpers/esm/interopRequireDefault.js",
-    "./helpers/interopRequireWildcard": [
-      {
-        "node": "./helpers/interopRequireWildcard.js",
-        "import": "./helpers/esm/interopRequireWildcard.js",
-        "default": "./helpers/interopRequireWildcard.js"
-      },
-      "./helpers/interopRequireWildcard.js"
-    ],
-    "./helpers/esm/interopRequireWildcard": "./helpers/esm/interopRequireWildcard.js",
-    "./helpers/newArrowCheck": [
-      {
-        "node": "./helpers/newArrowCheck.js",
-        "import": "./helpers/esm/newArrowCheck.js",
-        "default": "./helpers/newArrowCheck.js"
-      },
-      "./helpers/newArrowCheck.js"
-    ],
-    "./helpers/esm/newArrowCheck": "./helpers/esm/newArrowCheck.js",
-    "./helpers/objectDestructuringEmpty": [
-      {
-        "node": "./helpers/objectDestructuringEmpty.js",
-        "import": "./helpers/esm/objectDestructuringEmpty.js",
-        "default": "./helpers/objectDestructuringEmpty.js"
-      },
-      "./helpers/objectDestructuringEmpty.js"
-    ],
-    "./helpers/esm/objectDestructuringEmpty": "./helpers/esm/objectDestructuringEmpty.js",
-    "./helpers/objectWithoutPropertiesLoose": [
-      {
-        "node": "./helpers/objectWithoutPropertiesLoose.js",
-        "import": "./helpers/esm/objectWithoutPropertiesLoose.js",
-        "default": "./helpers/objectWithoutPropertiesLoose.js"
-      },
-      "./helpers/objectWithoutPropertiesLoose.js"
-    ],
-    "./helpers/esm/objectWithoutPropertiesLoose": "./helpers/esm/objectWithoutPropertiesLoose.js",
-    "./helpers/objectWithoutProperties": [
-      {
-        "node": "./helpers/objectWithoutProperties.js",
-        "import": "./helpers/esm/objectWithoutProperties.js",
-        "default": "./helpers/objectWithoutProperties.js"
-      },
-      "./helpers/objectWithoutProperties.js"
-    ],
-    "./helpers/esm/objectWithoutProperties": "./helpers/esm/objectWithoutProperties.js",
-    "./helpers/assertThisInitialized": [
-      {
-        "node": "./helpers/assertThisInitialized.js",
-        "import": "./helpers/esm/assertThisInitialized.js",
-        "default": "./helpers/assertThisInitialized.js"
-      },
-      "./helpers/assertThisInitialized.js"
-    ],
-    "./helpers/esm/assertThisInitialized": "./helpers/esm/assertThisInitialized.js",
-    "./helpers/possibleConstructorReturn": [
-      {
-        "node": "./helpers/possibleConstructorReturn.js",
-        "import": "./helpers/esm/possibleConstructorReturn.js",
-        "default": "./helpers/possibleConstructorReturn.js"
-      },
-      "./helpers/possibleConstructorReturn.js"
-    ],
-    "./helpers/esm/possibleConstructorReturn": "./helpers/esm/possibleConstructorReturn.js",
-    "./helpers/createSuper": [
-      {
-        "node": "./helpers/createSuper.js",
-        "import": "./helpers/esm/createSuper.js",
-        "default": "./helpers/createSuper.js"
-      },
-      "./helpers/createSuper.js"
-    ],
-    "./helpers/esm/createSuper": "./helpers/esm/createSuper.js",
-    "./helpers/superPropBase": [
-      {
-        "node": "./helpers/superPropBase.js",
-        "import": "./helpers/esm/superPropBase.js",
-        "default": "./helpers/superPropBase.js"
-      },
-      "./helpers/superPropBase.js"
-    ],
-    "./helpers/esm/superPropBase": "./helpers/esm/superPropBase.js",
-    "./helpers/get": [
-      {
-        "node": "./helpers/get.js",
-        "import": "./helpers/esm/get.js",
-        "default": "./helpers/get.js"
-      },
-      "./helpers/get.js"
-    ],
-    "./helpers/esm/get": "./helpers/esm/get.js",
-    "./helpers/set": [
-      {
-        "node": "./helpers/set.js",
-        "import": "./helpers/esm/set.js",
-        "default": "./helpers/set.js"
-      },
-      "./helpers/set.js"
-    ],
-    "./helpers/esm/set": "./helpers/esm/set.js",
-    "./helpers/taggedTemplateLiteral": [
-      {
-        "node": "./helpers/taggedTemplateLiteral.js",
-        "import": "./helpers/esm/taggedTemplateLiteral.js",
-        "default": "./helpers/taggedTemplateLiteral.js"
-      },
-      "./helpers/taggedTemplateLiteral.js"
-    ],
-    "./helpers/esm/taggedTemplateLiteral": "./helpers/esm/taggedTemplateLiteral.js",
-    "./helpers/taggedTemplateLiteralLoose": [
-      {
-        "node": "./helpers/taggedTemplateLiteralLoose.js",
-        "import": "./helpers/esm/taggedTemplateLiteralLoose.js",
-        "default": "./helpers/taggedTemplateLiteralLoose.js"
-      },
-      "./helpers/taggedTemplateLiteralLoose.js"
-    ],
-    "./helpers/esm/taggedTemplateLiteralLoose": "./helpers/esm/taggedTemplateLiteralLoose.js",
-    "./helpers/readOnlyError": [
-      {
-        "node": "./helpers/readOnlyError.js",
-        "import": "./helpers/esm/readOnlyError.js",
-        "default": "./helpers/readOnlyError.js"
-      },
-      "./helpers/readOnlyError.js"
-    ],
-    "./helpers/esm/readOnlyError": "./helpers/esm/readOnlyError.js",
-    "./helpers/writeOnlyError": [
-      {
-        "node": "./helpers/writeOnlyError.js",
-        "import": "./helpers/esm/writeOnlyError.js",
-        "default": "./helpers/writeOnlyError.js"
-      },
-      "./helpers/writeOnlyError.js"
-    ],
-    "./helpers/esm/writeOnlyError": "./helpers/esm/writeOnlyError.js",
-    "./helpers/classNameTDZError": [
-      {
-        "node": "./helpers/classNameTDZError.js",
-        "import": "./helpers/esm/classNameTDZError.js",
-        "default": "./helpers/classNameTDZError.js"
-      },
-      "./helpers/classNameTDZError.js"
-    ],
-    "./helpers/esm/classNameTDZError": "./helpers/esm/classNameTDZError.js",
-    "./helpers/temporalUndefined": [
-      {
-        "node": "./helpers/temporalUndefined.js",
-        "import": "./helpers/esm/temporalUndefined.js",
-        "default": "./helpers/temporalUndefined.js"
-      },
-      "./helpers/temporalUndefined.js"
-    ],
-    "./helpers/esm/temporalUndefined": "./helpers/esm/temporalUndefined.js",
-    "./helpers/tdz": [
-      {
-        "node": "./helpers/tdz.js",
-        "import": "./helpers/esm/tdz.js",
-        "default": "./helpers/tdz.js"
-      },
-      "./helpers/tdz.js"
-    ],
-    "./helpers/esm/tdz": "./helpers/esm/tdz.js",
-    "./helpers/temporalRef": [
-      {
-        "node": "./helpers/temporalRef.js",
-        "import": "./helpers/esm/temporalRef.js",
-        "default": "./helpers/temporalRef.js"
-      },
-      "./helpers/temporalRef.js"
-    ],
-    "./helpers/esm/temporalRef": "./helpers/esm/temporalRef.js",
-    "./helpers/slicedToArray": [
-      {
-        "node": "./helpers/slicedToArray.js",
-        "import": "./helpers/esm/slicedToArray.js",
-        "default": "./helpers/slicedToArray.js"
-      },
-      "./helpers/slicedToArray.js"
-    ],
-    "./helpers/esm/slicedToArray": "./helpers/esm/slicedToArray.js",
-    "./helpers/slicedToArrayLoose": [
-      {
-        "node": "./helpers/slicedToArrayLoose.js",
-        "import": "./helpers/esm/slicedToArrayLoose.js",
-        "default": "./helpers/slicedToArrayLoose.js"
-      },
-      "./helpers/slicedToArrayLoose.js"
-    ],
-    "./helpers/esm/slicedToArrayLoose": "./helpers/esm/slicedToArrayLoose.js",
-    "./helpers/toArray": [
-      {
-        "node": "./helpers/toArray.js",
-        "import": "./helpers/esm/toArray.js",
-        "default": "./helpers/toArray.js"
-      },
-      "./helpers/toArray.js"
-    ],
-    "./helpers/esm/toArray": "./helpers/esm/toArray.js",
-    "./helpers/toConsumableArray": [
-      {
-        "node": "./helpers/toConsumableArray.js",
-        "import": "./helpers/esm/toConsumableArray.js",
-        "default": "./helpers/toConsumableArray.js"
-      },
-      "./helpers/toConsumableArray.js"
-    ],
-    "./helpers/esm/toConsumableArray": "./helpers/esm/toConsumableArray.js",
-    "./helpers/arrayWithoutHoles": [
-      {
-        "node": "./helpers/arrayWithoutHoles.js",
-        "import": "./helpers/esm/arrayWithoutHoles.js",
-        "default": "./helpers/arrayWithoutHoles.js"
-      },
-      "./helpers/arrayWithoutHoles.js"
-    ],
-    "./helpers/esm/arrayWithoutHoles": "./helpers/esm/arrayWithoutHoles.js",
-    "./helpers/arrayWithHoles": [
-      {
-        "node": "./helpers/arrayWithHoles.js",
-        "import": "./helpers/esm/arrayWithHoles.js",
-        "default": "./helpers/arrayWithHoles.js"
-      },
-      "./helpers/arrayWithHoles.js"
-    ],
-    "./helpers/esm/arrayWithHoles": "./helpers/esm/arrayWithHoles.js",
-    "./helpers/maybeArrayLike": [
-      {
-        "node": "./helpers/maybeArrayLike.js",
-        "import": "./helpers/esm/maybeArrayLike.js",
-        "default": "./helpers/maybeArrayLike.js"
-      },
-      "./helpers/maybeArrayLike.js"
-    ],
-    "./helpers/esm/maybeArrayLike": "./helpers/esm/maybeArrayLike.js",
-    "./helpers/iterableToArray": [
-      {
-        "node": "./helpers/iterableToArray.js",
-        "import": "./helpers/esm/iterableToArray.js",
-        "default": "./helpers/iterableToArray.js"
-      },
-      "./helpers/iterableToArray.js"
-    ],
-    "./helpers/esm/iterableToArray": "./helpers/esm/iterableToArray.js",
-    "./helpers/unsupportedIterableToArray": [
-      {
-        "node": "./helpers/unsupportedIterableToArray.js",
-        "import": "./helpers/esm/unsupportedIterableToArray.js",
-        "default": "./helpers/unsupportedIterableToArray.js"
-      },
-      "./helpers/unsupportedIterableToArray.js"
-    ],
-    "./helpers/esm/unsupportedIterableToArray": "./helpers/esm/unsupportedIterableToArray.js",
-    "./helpers/arrayLikeToArray": [
-      {
-        "node": "./helpers/arrayLikeToArray.js",
-        "import": "./helpers/esm/arrayLikeToArray.js",
-        "default": "./helpers/arrayLikeToArray.js"
-      },
-      "./helpers/arrayLikeToArray.js"
-    ],
-    "./helpers/esm/arrayLikeToArray": "./helpers/esm/arrayLikeToArray.js",
-    "./helpers/nonIterableSpread": [
-      {
-        "node": "./helpers/nonIterableSpread.js",
-        "import": "./helpers/esm/nonIterableSpread.js",
-        "default": "./helpers/nonIterableSpread.js"
-      },
-      "./helpers/nonIterableSpread.js"
-    ],
-    "./helpers/esm/nonIterableSpread": "./helpers/esm/nonIterableSpread.js",
-    "./helpers/nonIterableRest": [
-      {
-        "node": "./helpers/nonIterableRest.js",
-        "import": "./helpers/esm/nonIterableRest.js",
-        "default": "./helpers/nonIterableRest.js"
-      },
-      "./helpers/nonIterableRest.js"
-    ],
-    "./helpers/esm/nonIterableRest": "./helpers/esm/nonIterableRest.js",
-    "./helpers/createForOfIteratorHelper": [
-      {
-        "node": "./helpers/createForOfIteratorHelper.js",
-        "import": "./helpers/esm/createForOfIteratorHelper.js",
-        "default": "./helpers/createForOfIteratorHelper.js"
-      },
-      "./helpers/createForOfIteratorHelper.js"
-    ],
-    "./helpers/esm/createForOfIteratorHelper": "./helpers/esm/createForOfIteratorHelper.js",
-    "./helpers/createForOfIteratorHelperLoose": [
-      {
-        "node": "./helpers/createForOfIteratorHelperLoose.js",
-        "import": "./helpers/esm/createForOfIteratorHelperLoose.js",
-        "default": "./helpers/createForOfIteratorHelperLoose.js"
-      },
-      "./helpers/createForOfIteratorHelperLoose.js"
-    ],
-    "./helpers/esm/createForOfIteratorHelperLoose": "./helpers/esm/createForOfIteratorHelperLoose.js",
-    "./helpers/skipFirstGeneratorNext": [
-      {
-        "node": "./helpers/skipFirstGeneratorNext.js",
-        "import": "./helpers/esm/skipFirstGeneratorNext.js",
-        "default": "./helpers/skipFirstGeneratorNext.js"
-      },
-      "./helpers/skipFirstGeneratorNext.js"
-    ],
-    "./helpers/esm/skipFirstGeneratorNext": "./helpers/esm/skipFirstGeneratorNext.js",
-    "./helpers/toPrimitive": [
-      {
-        "node": "./helpers/toPrimitive.js",
-        "import": "./helpers/esm/toPrimitive.js",
-        "default": "./helpers/toPrimitive.js"
-      },
-      "./helpers/toPrimitive.js"
-    ],
-    "./helpers/esm/toPrimitive": "./helpers/esm/toPrimitive.js",
-    "./helpers/toPropertyKey": [
-      {
-        "node": "./helpers/toPropertyKey.js",
-        "import": "./helpers/esm/toPropertyKey.js",
-        "default": "./helpers/toPropertyKey.js"
-      },
-      "./helpers/toPropertyKey.js"
-    ],
-    "./helpers/esm/toPropertyKey": "./helpers/esm/toPropertyKey.js",
-    "./helpers/initializerWarningHelper": [
-      {
-        "node": "./helpers/initializerWarningHelper.js",
-        "import": "./helpers/esm/initializerWarningHelper.js",
-        "default": "./helpers/initializerWarningHelper.js"
-      },
-      "./helpers/initializerWarningHelper.js"
-    ],
-    "./helpers/esm/initializerWarningHelper": "./helpers/esm/initializerWarningHelper.js",
-    "./helpers/initializerDefineProperty": [
-      {
-        "node": "./helpers/initializerDefineProperty.js",
-        "import": "./helpers/esm/initializerDefineProperty.js",
-        "default": "./helpers/initializerDefineProperty.js"
-      },
-      "./helpers/initializerDefineProperty.js"
-    ],
-    "./helpers/esm/initializerDefineProperty": "./helpers/esm/initializerDefineProperty.js",
-    "./helpers/applyDecoratedDescriptor": [
-      {
-        "node": "./helpers/applyDecoratedDescriptor.js",
-        "import": "./helpers/esm/applyDecoratedDescriptor.js",
-        "default": "./helpers/applyDecoratedDescriptor.js"
-      },
-      "./helpers/applyDecoratedDescriptor.js"
-    ],
-    "./helpers/esm/applyDecoratedDescriptor": "./helpers/esm/applyDecoratedDescriptor.js",
-    "./helpers/classPrivateFieldLooseKey": [
-      {
-        "node": "./helpers/classPrivateFieldLooseKey.js",
-        "import": "./helpers/esm/classPrivateFieldLooseKey.js",
-        "default": "./helpers/classPrivateFieldLooseKey.js"
-      },
-      "./helpers/classPrivateFieldLooseKey.js"
-    ],
-    "./helpers/esm/classPrivateFieldLooseKey": "./helpers/esm/classPrivateFieldLooseKey.js",
-    "./helpers/classPrivateFieldLooseBase": [
-      {
-        "node": "./helpers/classPrivateFieldLooseBase.js",
-        "import": "./helpers/esm/classPrivateFieldLooseBase.js",
-        "default": "./helpers/classPrivateFieldLooseBase.js"
-      },
-      "./helpers/classPrivateFieldLooseBase.js"
-    ],
-    "./helpers/esm/classPrivateFieldLooseBase": "./helpers/esm/classPrivateFieldLooseBase.js",
-    "./helpers/classPrivateFieldGet": [
-      {
-        "node": "./helpers/classPrivateFieldGet.js",
-        "import": "./helpers/esm/classPrivateFieldGet.js",
-        "default": "./helpers/classPrivateFieldGet.js"
-      },
-      "./helpers/classPrivateFieldGet.js"
-    ],
-    "./helpers/esm/classPrivateFieldGet": "./helpers/esm/classPrivateFieldGet.js",
-    "./helpers/classPrivateFieldSet": [
-      {
-        "node": "./helpers/classPrivateFieldSet.js",
-        "import": "./helpers/esm/classPrivateFieldSet.js",
-        "default": "./helpers/classPrivateFieldSet.js"
-      },
-      "./helpers/classPrivateFieldSet.js"
-    ],
-    "./helpers/esm/classPrivateFieldSet": "./helpers/esm/classPrivateFieldSet.js",
-    "./helpers/classPrivateFieldDestructureSet": [
-      {
-        "node": "./helpers/classPrivateFieldDestructureSet.js",
-        "import": "./helpers/esm/classPrivateFieldDestructureSet.js",
-        "default": "./helpers/classPrivateFieldDestructureSet.js"
-      },
-      "./helpers/classPrivateFieldDestructureSet.js"
-    ],
-    "./helpers/esm/classPrivateFieldDestructureSet": "./helpers/esm/classPrivateFieldDestructureSet.js",
-    "./helpers/classExtractFieldDescriptor": [
-      {
-        "node": "./helpers/classExtractFieldDescriptor.js",
-        "import": "./helpers/esm/classExtractFieldDescriptor.js",
-        "default": "./helpers/classExtractFieldDescriptor.js"
-      },
-      "./helpers/classExtractFieldDescriptor.js"
-    ],
-    "./helpers/esm/classExtractFieldDescriptor": "./helpers/esm/classExtractFieldDescriptor.js",
-    "./helpers/classStaticPrivateFieldSpecGet": [
-      {
-        "node": "./helpers/classStaticPrivateFieldSpecGet.js",
-        "import": "./helpers/esm/classStaticPrivateFieldSpecGet.js",
-        "default": "./helpers/classStaticPrivateFieldSpecGet.js"
-      },
-      "./helpers/classStaticPrivateFieldSpecGet.js"
-    ],
-    "./helpers/esm/classStaticPrivateFieldSpecGet": "./helpers/esm/classStaticPrivateFieldSpecGet.js",
-    "./helpers/classStaticPrivateFieldSpecSet": [
-      {
-        "node": "./helpers/classStaticPrivateFieldSpecSet.js",
-        "import": "./helpers/esm/classStaticPrivateFieldSpecSet.js",
-        "default": "./helpers/classStaticPrivateFieldSpecSet.js"
-      },
-      "./helpers/classStaticPrivateFieldSpecSet.js"
-    ],
-    "./helpers/esm/classStaticPrivateFieldSpecSet": "./helpers/esm/classStaticPrivateFieldSpecSet.js",
-    "./helpers/classStaticPrivateMethodGet": [
-      {
-        "node": "./helpers/classStaticPrivateMethodGet.js",
-        "import": "./helpers/esm/classStaticPrivateMethodGet.js",
-        "default": "./helpers/classStaticPrivateMethodGet.js"
-      },
-      "./helpers/classStaticPrivateMethodGet.js"
-    ],
-    "./helpers/esm/classStaticPrivateMethodGet": "./helpers/esm/classStaticPrivateMethodGet.js",
-    "./helpers/classStaticPrivateMethodSet": [
-      {
-        "node": "./helpers/classStaticPrivateMethodSet.js",
-        "import": "./helpers/esm/classStaticPrivateMethodSet.js",
-        "default": "./helpers/classStaticPrivateMethodSet.js"
-      },
-      "./helpers/classStaticPrivateMethodSet.js"
-    ],
-    "./helpers/esm/classStaticPrivateMethodSet": "./helpers/esm/classStaticPrivateMethodSet.js",
+    "./helpers/esm/classApplyDescriptorDestructureSet": "./helpers/esm/classApplyDescriptorDestructureSet.js",
     "./helpers/classApplyDescriptorGet": [
       {
         "node": "./helpers/classApplyDescriptorGet.js",
