@@ -1,6 +1,6 @@
 {
   "name": "@emailjs/browser",
-  "version": "3.11.0",
+  "version": "4.4.1",
   "description": "Official EmailJS SDK for browsers",
   "private": false,
   "author": "EmailJS",
@@ -23,13 +23,14 @@
     "node": ">=14.0.0"
   },
   "scripts": {
-    "_clean": "rm -rf es && rm -rf cjs && rm -rf dist",
-    "_set-version": "./set-version.sh $npm_package_version",
-    "_build-ts": "tsc --declaration && npm run _set-version es",
-    "_build-cjs": "tsc --module commonjs --outDir cjs && npm run _set-version cjs",
-    "_build-bundle": "webpack --env production",
-    "build": "npm run _clean && npm run _build-ts && npm run _build-cjs && npm run _build-bundle",
-    "test": "jest --coverage",
+    "build:clean": "rm -rf es && rm -rf cjs && rm -rf dist",
+    "build:version": "./set-version.sh $npm_package_version",
+    "build:ts": "tsc --declaration --project ./tsconfig.build.json && npm run build:version es",
+    "build:cjs": "tsc --module commonjs --outDir cjs --project ./tsconfig.build.json && npm run build:version cjs",
+    "build:lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "build:bundle": "rollup -c",
+    "build": "npm run build:clean && npm run build:lint && npm run build:ts && npm run build:cjs && npm run build:bundle",
+    "test": "jest --coverage --no-cache",
     "lint": "tsc --noEmit && eslint src"
   },
   "keywords": [
@@ -45,24 +46,20 @@
     "emailjs-com"
   ],
   "devDependencies": {
-    "@babel/preset-env": "^7.21.4",
-    "@babel/preset-typescript": "^7.21.4",
-    "@types/jest": "^29.5.0",
-    "@typescript-eslint/eslint-plugin": "^5.57.1",
-    "@typescript-eslint/parser": "^5.57.1",
-    "babel-jest": "^29.5.0",
-    "babel-loader": "^9.1.2",
-    "core-js": "^3.30.0",
-    "eslint": "^8.37.0",
-    "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "^2.27.5",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "prettier": "^2.8.7",
-    "terser-webpack-plugin": "^5.3.7",
-    "typescript": "^5.0.3",
-    "webpack": "^5.78.0",
-    "webpack-cli": "^5.0.1"
+    "@babel/preset-env": "^7.24.8",
+    "@babel/preset-typescript": "^7.24.7",
+    "@happy-dom/jest-environment": "^14.12.3",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@typescript-eslint/eslint-plugin": "^7.16.0",
+    "@typescript-eslint/parser": "^7.16.0",
+    "core-js": "^3.37.1",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "jest": "^29.7.0",
+    "prettier": "^3.3.2",
+    "rollup": "^4.18.1",
+    "typescript": "^5.5.3"
   },
   "prettier": {
     "trailingComma": "all",
