@@ -1,23 +1,25 @@
 {
 	"name": "ansi-styles",
-	"version": "3.2.1",
+	"version": "4.3.0",
 	"description": "ANSI escape codes for styling strings in the terminal",
 	"license": "MIT",
 	"repository": "chalk/ansi-styles",
+	"funding": "https://github.com/chalk/ansi-styles?sponsor=1",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
 		"url": "sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=4"
+		"node": ">=8"
 	},
 	"scripts": {
-		"test": "xo && ava",
+		"test": "xo && ava && tsd",
 		"screenshot": "svg-term --command='node screenshot' --out=screenshot.svg --padding=3 --width=55 --height=3 --at=1000 --no-cursor"
 	},
 	"files": [
-		"index.js"
+		"index.js",
+		"index.d.ts"
 	],
 	"keywords": [
 		"ansi",
@@ -42,15 +44,13 @@
 		"text"
 	],
 	"dependencies": {
-		"color-convert": "^1.9.0"
+		"color-convert": "^2.0.1"
 	},
 	"devDependencies": {
-		"ava": "*",
-		"babel-polyfill": "^6.23.0",
+		"@types/color-convert": "^1.9.0",
+		"ava": "^2.3.0",
 		"svg-term-cli": "^2.1.1",
-		"xo": "*"
-	},
-	"ava": {
-		"require": "babel-polyfill"
+		"tsd": "^0.11.0",
+		"xo": "^0.25.3"
 	}
 }
