@@ -1,10 +1,9 @@
 {
   "name": "camera-controls",
-  "version": "2.7.3",
+  "version": "2.9.0",
   "author": "Yomotsu",
   "license": "MIT",
-  "main": "dist/camera-controls.js",
-  "jsnext:main": "dist/camera-controls.module.js",
+  "main": "dist/camera-controls.cjs",
   "module": "dist/camera-controls.module.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -12,26 +11,27 @@
   ],
   "repository": "yomotsu/camera-controls",
   "devDependencies": {
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-typescript": "^11.1.3",
-    "@types/three": "^0.155.1",
-    "@typescript-eslint/eslint-plugin": "^6.5.0",
-    "@typescript-eslint/parser": "^6.5.0",
+    "@rollup/plugin-replace": "^5.0.7",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@stylistic/eslint-plugin": "^2.6.4",
+    "@types/three": "^0.167.1",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
     "downlevel-dts": "^0.11.0",
-    "eslint": "^8.48.0",
+    "eslint": "^9.9.0",
     "eslint-config-mdcs": "^5.0.0",
-    "husky": "^8.0.3",
-    "lint-staged": "^14.0.1",
+    "husky": "^9.1.4",
+    "lint-staged": "^15.2.9",
     "npm-run-all": "^4.1.5",
-    "open-cli": "^7.2.0",
-    "rollup": "^3.28.1",
-    "semantic-release": "^21.1.1",
-    "serve": "^14.2.1",
-    "terser": "^5.19.3",
-    "three": "^0.155.0",
-    "tslib": "^2.6.2",
-    "typedoc": "^0.25.0",
-    "typescript": "^5.2.2"
+    "open-cli": "^8.0.0",
+    "rollup": "^4.20.0",
+    "semantic-release": "^23.1.1",
+    "serve": "^14.2.3",
+    "terser": "^5.31.6",
+    "three": "^0.167.1",
+    "tslib": "^2.6.3",
+    "typedoc": "^0.26.5",
+    "typescript": "^5.5.4"
   },
   "peerDependencies": {
     "three": ">=0.126.1"
@@ -42,7 +42,7 @@
     "dev:rollup": "rollup --config --watch",
     "dev:serve": "serve -S -p 3000 ./ && kill $!",
     "dev:open": "open-cli http://localhost:3000/examples/",
-    "build": "rollup --config && terser dist/camera-controls.js -o dist/camera-controls.min.js --comments '/^!/' && downlevel-dts . .",
+    "build": "rollup --config && terser dist/camera-controls.module.js -o dist/camera-controls.module.min.js --comments '/^!/' && downlevel-dts . .",
     "prepack": "npm run build",
     "lint": "eslint src --ext ts",
     "typedoc": "typedoc",
@@ -58,13 +58,14 @@
     "parser": "@typescript-eslint/parser",
     "extends": "mdcs",
     "plugins": [
+      "@stylistic/eslint-plugin",
       "@typescript-eslint"
     ],
     "rules": {
       "no-unused-vars": 0,
       "@typescript-eslint/no-unused-vars": 1,
       "indent": 0,
-      "@typescript-eslint/indent": [
+      "@stylistic/indent": [
         "error",
         "tab",
         {
