@@ -1,22 +1,21 @@
 {
 	"name": "chalk",
-	"version": "2.4.2",
+	"version": "4.1.2",
 	"description": "Terminal string styling done right",
 	"license": "MIT",
 	"repository": "chalk/chalk",
+	"funding": "https://github.com/chalk/chalk?sponsor=1",
+	"main": "source",
 	"engines": {
-		"node": ">=4"
+		"node": ">=10"
 	},
 	"scripts": {
-		"test": "xo && tsc --project types && flow --max-warnings=0 && nyc ava",
-		"bench": "matcha benchmark.js",
-		"coveralls": "nyc report --reporter=text-lcov | coveralls"
+		"test": "xo && nyc ava && tsd",
+		"bench": "matcha benchmark.js"
 	},
 	"files": [
-		"index.js",
-		"templates.js",
-		"types/index.d.ts",
-		"index.js.flow"
+		"source",
+		"index.d.ts"
 	],
 	"keywords": [
 		"color",
@@ -42,30 +41,28 @@
 		"text"
 	],
 	"dependencies": {
-		"ansi-styles": "^3.2.1",
-		"escape-string-regexp": "^1.0.5",
-		"supports-color": "^5.3.0"
+		"ansi-styles": "^4.1.0",
+		"supports-color": "^7.1.0"
 	},
 	"devDependencies": {
-		"ava": "*",
-		"coveralls": "^3.0.0",
-		"execa": "^0.9.0",
-		"flow-bin": "^0.68.0",
-		"import-fresh": "^2.0.0",
+		"ava": "^2.4.0",
+		"coveralls": "^3.0.7",
+		"execa": "^4.0.0",
+		"import-fresh": "^3.1.0",
 		"matcha": "^0.7.0",
-		"nyc": "^11.0.2",
-		"resolve-from": "^4.0.0",
-		"typescript": "^2.5.3",
-		"xo": "*"
+		"nyc": "^15.0.0",
+		"resolve-from": "^5.0.0",
+		"tsd": "^0.7.4",
+		"xo": "^0.28.2"
 	},
-	"types": "types/index.d.ts",
 	"xo": {
-		"envs": [
-			"node",
-			"mocha"
-		],
-		"ignores": [
-			"test/_flow.js"
-		]
+		"rules": {
+			"unicorn/prefer-string-slice": "off",
+			"unicorn/prefer-includes": "off",
+			"@typescript-eslint/member-ordering": "off",
+			"no-redeclare": "off",
+			"unicorn/string-content": "off",
+			"unicorn/better-regex": "off"
+		}
 	}
 }
