@@ -1,6 +1,6 @@
 {
   "name": "convert-source-map",
-  "version": "1.9.0",
+  "version": "2.0.0",
   "description": "Converts a source-map from/to  different formats and allows adding/changing properties.",
   "main": "index.js",
   "scripts": {
@@ -30,12 +30,9 @@
   },
   "license": "MIT",
   "engine": {
-    "node": ">=0.6"
+    "node": ">=4"
   },
   "files": [
     "index.js"
-  ],
-  "browser": {
-    "fs": false
-  }
+  ]
 }
