@@ -1,11 +1,7 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
-// https://vitejs.dev/config/
+// https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
-    build: {
-        chunkSizeWarningLimit: 100000000
-    },
-    base: "/3d-react-portfolio",
 })
